Show structural checkbox sincronized with Key_0 onkeypress event
