+++ tractome.py
@@ -53,7 +53,7 @@ class Tractome(object):
         self.list_oper_ROIS = []
      
      
-    def loading_structural(self, structpath = None):
+    def loading_structural(self, structpath = None,  actor_name = 'Volume Slicer'):
         """
         Loading structural data.
         """
@@ -71,14 +71,15 @@ class Tractome(object):
             data = data1
             del data1
             
+                   
         
         # Create the Guillotine object
         data = (np.interp(data, [data.min(), data.max()], [0, 255]))
-        self.guil = Guillotine('Volume Slicer', data, np.copy(self.affine))
+        self.guil = Guillotine(actor_name, data, np.copy(self.affine))
         self.scene.add_actor(self.guil) 
     
         
-    def loading_full_tractograpy(self, tracpath=None):
+    def loading_full_tractograpy(self, tracpath=None,  actor_name = 'Bundle Picker'):
         """
         Loading full tractography and creates StreamlineLabeler to
         show it all.
@@ -124,7 +125,7 @@ class Tractome(object):
             self.update_info(general_info_filename)
                     
         # create the interaction system for tracks, 
-        self.streamlab  = StreamlineLabeler('Bundle Picker',
+        self.streamlab  = StreamlineLabeler(actor_name,
                                             self.buffers, self.clusters,
                                             vol_shape=self.dims, 
                                             affine=np.copy(self.affine),
